Add player spawn direction vector
