+++ assets/maps/invalid/rgb/missing.cub
@@ -1,0 +1,6 @@
+NO ./assets/textures/test.xpm
+SO ./assets/textures/test.xpm
+WE ./assets/textures/test.xpm
+EA ./assets/textures/test.xpm
+
+F 255,0,99
