--- assets/maps/invalid/rgb/missing_values.cub
@@ -1,17 +1,0 @@
-NO ./assets/textures/test.xpm
-SO ./assets/textures/test.xpm
-EA ./assets/textures/test.xpm
-WE ./assets/textures/test.xpm
-
-F 210,210
-C 210,210,30
-
-    111111111
-    1E0000001
-    11111000111
-    10001000001
-111110001000001
-100000001000001
-100111000011111
-1001 100001
-1111 111111
