+++ assets/maps/invalid/tiles/unknown_char.cub
