+++ src/colour.c
@@ -1,0 +1,18 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   colour.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: lmells <lmells@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/09/07 12:35:00 by lmells            #+#    #+#             */
+/*   Updated: 2023/09/07 13:12:14 by lmells           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include <cub3d.h>
+
+uint64_t	rgb_to_uint64(uint8_t r, uint8_t g, uint8_t b)
+{
+	return (r << 16 | g << 8 | b);
+}
