+++ src/parser_rgb.c
